Create net standard project.
