+++ StringTokenFormatter.Shared/DefaultTokenMarkers.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace AndyWilsonUk.StringTokenFormatter
+{
+	public class DefaultTokenMarkers : TokenMarkers
+    {
+        public string StartToken
+        {
+            get { return "{"; }
+        }
+
+        public string EndToken
+        {
+            get { return "}"; }
+        }
+
+        public string StartTokenEscaped
+        {
+            get { return StartToken + StartToken; }
+        }
+
+        public IEqualityComparer<string> TokenNameComparer
+        {
+            get { return StringComparer.InvariantCultureIgnoreCase; }
+        }
+    }
+}
