+++ StringTokenFormatter.Shared/TokenMarkers.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+namespace AndyWilsonUk.StringTokenFormatter
+{
+    public interface TokenMarkers
+    {
+        string StartToken { get; }
+        string EndToken { get; }
+        string StartTokenEscaped { get; }
+        IEqualityComparer<string> TokenNameComparer { get; }
+    }
+}
