+++ StringTokenFormatter.Tests/StringTokenExtensionsTests.cs
@@ -1,12 +1,11 @@
-﻿using System;
-using System.Text;
+﻿using AndyWilsonUk.StringTokenFormatter;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Globalization;
 
 namespace StringTokenFormatter.Tests
 {
-    [TestClass]
+	[TestClass]
     public class StringTokenExtensionsTests
     {
         [TestMethod]
