+++ StringTokenFormatter.Tests/TokenReplacerFormatTests.cs
@@ -1,12 +1,11 @@
-﻿using System;
+﻿using AndyWilsonUk.StringTokenFormatter;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
-using StringTokenFormatter;
 using System.Globalization;
 
 namespace StringTokenFormatter.Tests
 {
-    [TestClass]
+	[TestClass]
     public class TokenReplacerFormatTests
     {
         [TestMethod]
