+++ StringTokenFormatter.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30110.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2005
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StringTokenFormatter", "StringTokenFormatter\StringTokenFormatter.csproj", "{F4B7425A-2A79-4665-A58F-6BCE3E9C1730}"
 EndProject
@@ -9,7 +9,15 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StringTokenFormatter.Tests"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StringTokenFormatter.Packager", "StringTokenFormatter.Packager\StringTokenFormatter.Packager.csproj", "{0451BAEF-DF2E-4B98-8644-94EE9415E389}"
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "StringTokenFormatter.Shared", "StringTokenFormatter.Shared\StringTokenFormatter.Shared.shproj", "{7C232228-E18A-47F6-B7DB-EDE10485E727}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StringTokenFormatter.NetStandard", "StringTokenFormatter.NetStandard\StringTokenFormatter.NetStandard.csproj", "{FBF29F48-A1C2-409E-8CAA-334168875FEA}"
+EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		StringTokenFormatter.Shared\StringTokenFormatter.Shared.projitems*{7c232228-e18a-47f6-b7db-ede10485e727}*SharedItemsImports = 13
+		StringTokenFormatter.Shared\StringTokenFormatter.Shared.projitems*{f4b7425a-2a79-4665-a58f-6bce3e9c1730}*SharedItemsImports = 4
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
@@ -27,8 +35,15 @@ Global
 		{0451BAEF-DF2E-4B98-8644-94EE9415E389}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0451BAEF-DF2E-4B98-8644-94EE9415E389}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0451BAEF-DF2E-4B98-8644-94EE9415E389}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FBF29F48-A1C2-409E-8CAA-334168875FEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FBF29F48-A1C2-409E-8CAA-334168875FEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FBF29F48-A1C2-409E-8CAA-334168875FEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FBF29F48-A1C2-409E-8CAA-334168875FEA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F198AB7F-2667-439D-87F3-8C7642EAB85B}
 	EndGlobalSection
 EndGlobal
