--- StringTokenFormatter/TokenMarkers.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-namespace StringTokenFormatter
-{
-    public interface TokenMarkers
-    {
-        string StartToken { get; }
-        string EndToken { get; }
-        string StartTokenEscaped { get; }
-        IEqualityComparer<string> TokenNameComparer { get; }
-    }
-}
